requirements.txt
Django
djangorestframework
numpy
olefile
opencv-python
Pillow
pkg-resources
pytz
